application-specific/src/agentic/requirements.txt
# Core Dependencies
langchain>=0.1.0
langchain-core>=0.1.0
langchain-openai>=0.0.5
langgraph>=0.0.40
langsmith>=0.0.70

# MCP Client
mcp>=0.1.0
httpx>=0.25.0
sse-starlette>=1.8.0

# UI Dependencies
PySide6>=6.6.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.0.0

# Optional: For enhanced features
# Visualization
matplotlib>=3.7.0
networkx>=3.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0
